+++ zarrs/src/array.rs
@@ -199,7 +199,7 @@ pub fn chunk_shape_to_array_shape(chunk_shape: &[std::num::NonZeroU64]) -> Array
 /// - Raw bytes variants: [`ArrayBytes`]
 /// - Typed element variants: e.g. `Vec<T>` where `T: Element`
 /// - `ndarray` variants: `ndarray::ArrayD<T>` where `T: Element` (requires `ndarray` feature)
-/// - `dlpack` variants: [`Tensor`], which is convertible into a `DLPack` managed tensor (requires `dlpack` feature)
+/// - `dlpack` variants: [`Tensor`], where a `Tensor<'static>` is convertible into a `DLPack` managed tensor (requires `dlpack` feature)
 ///
 /// Similarly, array `store_*` methods are generic over the input type.
 ///