+++ zarrs/src/array/into_array_bytes.rs
@@ -91,8 +91,8 @@ impl<T: Element, D: ndarray::Dimension> IntoArrayBytes<'static> for ndarray::Arr
     }
 }
 
-impl IntoArrayBytes<'static> for super::Tensor {
-    fn into_array_bytes(self, data_type: &DataType) -> Result<ArrayBytes<'static>, ElementError> {
+impl<'a> IntoArrayBytes<'a> for super::Tensor<'a> {
+    fn into_array_bytes(self, data_type: &DataType) -> Result<ArrayBytes<'a>, ElementError> {
         let (bytes, tensor_data_type, _) = self.into_parts();
         if tensor_data_type != *data_type {
             return Err(ElementError::IncompatibleElementType);
@@ -101,7 +101,7 @@ impl IntoArrayBytes<'static> for super::Tensor {
     }
 }
 
-impl<'a> IntoArrayBytes<'a> for &'a super::Tensor {
+impl<'a> IntoArrayBytes<'a> for &'a super::Tensor<'_> {
     fn into_array_bytes(self, data_type: &DataType) -> Result<ArrayBytes<'a>, ElementError> {
         let (bytes, tensor_data_type, _) = self.as_parts();
         if tensor_data_type != data_type {